requirements.txt
numpy==2.4.2
pandas==2.3.3
scipy==1.17.0
matplotlib==3.10.8
seaborn==0.13.2
plotly==6.5.2
PyYAML==6.0
